Make minified scripts work

+++ src/SoftinuxBase.Security/Scripts/security_user_permissions.js
@@ -6,10 +6,16 @@
 /// <reference path = '../../SoftinuxBase.Barebone/Scripts/barebone_ajax.js' />
 /// <reference path = './security_user.js' />
 
+let bareboneAjaxModule;
+try {
+    bareboneAjaxModule = await import('/Scripts.barebone.min.js');
+} catch (err) {
+    bareboneAjaxModule = await import('/Scripts.barebone_ajax.js');
+}
+
 'use strict';
 
-import makeAjaxRequest from '/Scripts/barebone_ajax.js';
-import { inputFormGroupSetError, inputFormGroupValidator } from '/Scripts/security_user.js';
+//import { inputFormGroupSetError, inputFormGroupValidator } from '/Scripts/security_user.js';
 import { inputOnlyAlphanumeric } from '/Scripts/toolbox.js';
 
 /* Select 2 Boostrap 4 Theme for all Select2
@@ -19,7 +25,7 @@ import { inputOnlyAlphanumeric } from '/Scripts/toolbox.js';
 
 // Select2: template for unlink icon in edit role selectbox
 function iformat(icon) {
-    var originalOption = icon.element;
+    const originalOption = icon.element;
     return $('<span><i class="' + $(originalOption).data('icon-base') + ' ' + $(originalOption).data('icon') + '"></i> ' + icon.text + '</span>');
 }
 
@@ -37,27 +43,44 @@ window.deleteRole = deleteRole;
 /* ------------------------ events handlers ----------------------- */
 /* ---------------------------------------------------------------- */
 
-document.getElementById('bulk-delete-btn').addEventListener('click', () => {
-    if (document.getElementById('availableRolesForDelete').selectedOptions.length === 0)
-        window.toastr.warning('No role selected', 'Warning');
-    else {
-        deleteRole(Array.from(document.getElementById('availableRolesForDelete').options).filter(option => option.selected).map(option => option.value));
-    }
-    console.log(Array.from(document.getElementById('availableRolesForDelete').options).filter(option => option.selected).map(option => option.value));
-});
+const bulkDeleteBtn = document.getElementById('bulk-delete-btn');
+if (bulkDeleteBtn) {
+    bulkDeleteBtn.addEventListener('click',
+        () => {
+            if (document.getElementById('availableRolesForDelete').selectedOptions.length === 0)
+                window.toastr.warning('No role selected', 'Warning');
+            else {
+                deleteRole(Array.from(document.getElementById('availableRolesForDelete').options)
+                    .filter(option => option.selected).map(option => option.value));
+            }
+            console.log(Array.from(document.getElementById('availableRolesForDelete').options)
+                .filter(option => option.selected).map(option => option.value));
+        });
+}
+const cancelBulkDeleteBtn = document.getElementById('cancel-bulk-delete-btn');
+if (cancelBulkDeleteBtn) {
+    cancelBulkDeleteBtn.addEventListener('click',
+        () => {
+            document.getElementById('availableRolesForDelete').selectedIndex = -1;
+        });
+}
 
-document.getElementById('cancel-bulk-delete-btn').addEventListener('click', () => {
-    document.getElementById('availableRolesForDelete').selectedIndex = -1;
-});
+const saveRenameRoleBtn = document.getElementById('save-rename-role-btn');
+if (saveRenameRoleBtn) {
+    saveRenameRoleBtn.addEventListener('click',
+        (event_) => {
+            saveEditRoleName();
+        });
+}
 
-document.getElementById('save-rename-role-btn').addEventListener('click', (event_) => {
-    saveEditRoleName();
-});
-
-document.getElementById('unlink-role-btn').addEventListener('click', (event_) => {
-    console.log(event_.target);
-    deleteAllPermissionsOfRole(event_.target.attributes['data-name']);
-});
+const unlinkRoleBtn = document.getElementById('unlink-role-btn');
+if (unlinkRoleBtn) {
+    unlinkRoleBtn.addEventListener('click',
+        (event_) => {
+            console.log(event_.target);
+            deleteAllPermissionsOfRole(event_.target.attributes['data-name']);
+        });
+}
 
 Array.prototype.forEach.call(document.querySelectorAll('select.update-role-permission'), function (element_) {
     $(element_).on('select2:select', function (event_) {
@@ -72,28 +95,43 @@ Array.prototype.forEach.call(document.querySelectorAll('select.update-role-permi
 });
 
 // Change
-document.getElementById('role_name_input').addEventListener('change', () => {
-    inputFormGroupValidator('#role_name_input');
-});
+const roleNameInput = document.getElementById('role_name_input');
+if (roleNameInput) {
+    roleNameInput.addEventListener('change',
+        () => {
+            inputFormGroupValidator('#role_name_input');
+        });
+}
 
 // Focusout
-document.getElementById('role_name_input').addEventListener('focusout', () => {
-    inputFormGroupValidator('#role_name_input');
-});
+if (roleNameInput) {
+    roleNameInput.addEventListener('focusout',
+        () => {
+            inputFormGroupValidator('#role_name_input');
+        });
+}
 
 // Keypress: only numbers
-document.getElementById('role_name_input').addEventListener('keypress', event_ => {
-    inputOnlyAlphanumeric(event_);
-}, false);
+if (roleNameInput) {
+    roleNameInput.addEventListener('keypress',
+        event_ => {
+            inputOnlyAlphanumeric(event_);
+        },
+        false);
+}
 
 /* -------------------------------------------------------------------------------------------- */
 /* ------------------------ Modal initialization ---------------------------------------------- */
 /* -------------------------------------------------------------------------------------------- */
 
-$('#renameRoleModal').on('show.bs.modal', function (event) {
-    const roleNameInputElt = document.getElementById('role_rename_input');
-    roleNameInputElt.value = '';
-});
+const renameRoleModal = $('#renameRoleModal');
+if (renameRoleModal.length) {
+    renameRoleModal.on('show.bs.modal',
+        function (event) {
+            const roleNameInputElt = document.getElementById('role_rename_input');
+            roleNameInputElt.value = '';
+        });
+}
 
 /* --------------------------------------------------------------------------------------------- */
 /* ------------------------ Exported functions ------------------------------------------------- */
@@ -108,36 +146,43 @@ $('#renameRoleModal').on('show.bs.modal', function (event) {
 /**
  * Save new role.
  */
-document.getElementById('save-add-role-btn').addEventListener('click', () => {
-    const roleNameInputElt = document.getElementById('role_name_input');
-    if (!roleNameInputElt.value) {
-        window.toastr.warning('No role name given.', 'Role not saved!');
-        inputFormGroupValidator('#role_name_input');
-        return;
-    }
+const saveAddRoleBtn = document.getElementById('save-add-role-btn');
+if (saveAddRoleBtn) {
+    saveAddRoleBtn.addEventListener('click',
+        () => {
+            const roleNameInputElt = document.getElementById('role_name_input');
+            if (!roleNameInputElt.value) {
+                window.toastr.warning('No role name given.', 'Role not saved!');
+                inputFormGroupValidator('#role_name_input');
+                return;
+            }
 
-    const postData = {
-        RoleName: roleNameInputElt.value
-    };
+            const postData = {
+                RoleName: roleNameInputElt.value
+            };
 
-    makeAjaxRequest('POST', '/administration/save-new-role', postData, (responseStatus_, responseText_) => {
-        if (responseStatus_ === 201) {
-            window.toastr.success(responseText_, 'New role created');
-            inputFormGroupSetError('#role_name_input', null);
-            refreshPermissionsTabs();
-            resetAddRoleForm();
-        } else {
-            inputFormGroupSetError('#role_name_input', responseText_ || responseStatus_);
-        }
-    });
-});
+            bareboneAjaxModule.makeAjaxRequest('POST',
+                '/administration/save-new-role',
+                postData,
+                (responseStatus_, responseText_) => {
+                    if (responseStatus_ === 201) {
+                        window.toastr.success(responseText_, 'New role created');
+                        inputFormGroupSetError('#role_name_input', null);
+                        refreshPermissionsTabs();
+                        resetAddRoleForm();
+                    } else {
+                        inputFormGroupSetError('#role_name_input', responseText_ || responseStatus_);
+                    }
+                });
+        });
+}
 
 /**
  * Update the UI with selected role information. Ajax GET.
  * @param {any} roleId_ - roleId
  */
 export function viewSelectedRole(roleId_) {
-    makeAjaxRequest('GET', '/administration/read-role', { roleId_: roleId_ }, (responseStatus_, responseText_) => {
+    bareboneAjaxModule.makeAjaxRequest('GET', '/administration/read-role', { roleId_: roleId_ }, (responseStatus_, responseText_) => {
         if (responseStatus_ !== 200) {
             window.toastr.error(responseText_, 'Error reading role');
             return;
@@ -207,7 +252,7 @@ function updateRolePermission(event_) {
         Add: event_.params.data.selected
     };
 
-    makeAjaxRequest('POST', '/administration/update-role-permission', params, (responseStatus_, responseText_) => {
+    bareboneAjaxModule.makeAjaxRequest('POST', '/administration/update-role-permission', params, (responseStatus_, responseText_) => {
         if (responseStatus_ === 204) {
             window.toastr.success(responseText_, 'Changes saved');
             // If this role's permissions are currently viewed, refresh
@@ -244,7 +289,7 @@ export function saveEditRoleName() {
         RoleName: roleNameInputElt.value
     };
 
-    makeAjaxRequest('POST', '/administration/update-role', postData, (responseStatus_, responseText_) => {
+    bareboneAjaxModule.makeAjaxRequest('POST', '/administration/update-role', postData, (responseStatus_, responseText_) => {
         if (responseStatus_ === 201) {
             window.toastr.success(responseText_, 'Changes saved');
             $('#renameRoleModal').modal('hide');
@@ -261,7 +306,7 @@ export function saveEditRoleName() {
 }
 
 export function deleteAllPermissionsOfRole(roleName_) {
-    makeAjaxRequest('DELETE', `/administration/remove-role-permissions/${roleName_}`, {}, (responseStatus_, responseText_) => {
+    bareboneAjaxModule.makeAjaxRequest('DELETE', `/administration/remove-role-permissions/${roleName_}`, {}, (responseStatus_, responseText_) => {
         if (responseStatus_ === 204) {
             window.toastr.success(`All permissions removed from role ${roleName_}`, 'Delete OK');
             refreshPermissionsTabs();
@@ -274,7 +319,7 @@ export function deleteAllPermissionsOfRole(roleName_) {
 }
 
 export function deleteRole(roleNameList_) {
-    makeAjaxRequest('DELETE', `/administration/delete-role/${roleNameList_}`, {}, (responseStatus_, responseText_) => {
+    bareboneAjaxModule.makeAjaxRequest('DELETE', `/administration/delete-role/${roleNameList_}`, {}, (responseStatus_, responseText_) => {
         if (responseStatus_ === 200) {
             window.toastr.success(`${roleNameList_}`, 'Role(s) deleted');
             refreshPermissionsTabs();
@@ -287,7 +332,7 @@ export function deleteRole(roleNameList_) {
 
 function reloadGrantPermissionsHtmlView() {
     return new window.Promise((resolve, reject) => {
-        makeAjaxRequest('GET', '/administration/read-permissions-grants', null, (responseStatus_, responseText_) => {
+        bareboneAjaxModule.makeAjaxRequest('GET', '/administration/read-permissions-grants', null, (responseStatus_, responseText_) => {
             document.getElementById('GrantPermissionsTable').innerHTML = responseText_;
             // Reattach the event listeners
             Array.prototype.forEach.call(document.querySelectorAll('select.update-role-permission'), function (element_) {
@@ -308,7 +353,7 @@ function reloadGrantPermissionsHtmlView() {
 
 function reloadEditRoleHtmlView() {
     return new window.Promise((resolve, reject) => {
-        makeAjaxRequest('GET', '/administration/edit-role-tab', null, (responseStatus_, responseText_) => {
+        bareboneAjaxModule.makeAjaxRequest('GET', '/administration/edit-role-tab', null, (responseStatus_, responseText_) => {
             document.getElementById('edit_role_form').innerHTML = responseText_;
             resolve();
         });
@@ -317,7 +362,7 @@ function reloadEditRoleHtmlView() {
 
 function reloadBulkDeleteTab() {
     return new window.Promise((resolve, reject) => {
-        makeAjaxRequest('GET', '/administration/bulk-delete-role-tab', null, (responseStatus_, responseText_) => {
+        bareboneAjaxModule.makeAjaxRequest('GET', '/administration/bulk-delete-role-tab', null, (responseStatus_, responseText_) => {
             document.getElementById('availableRolesForDelete').innerHTML = responseText_;
             resolve();
         });
@@ -328,8 +373,9 @@ function reloadBulkDeleteTab() {
 function refreshPermissionsTabs() {
     return Promise.all([reloadGrantPermissionsHtmlView(), reloadEditRoleHtmlView(), reloadBulkDeleteTab()])
         .then(() => {
-            document.getElementById('unlink-role-btn').addEventListener('click', () => {
-                deleteAllPermissionsOfRole(document.getElementById('edit_role_normalizedName').value);
+            let unlinkRoleBtn = document.getElementById('unlink-role-btn');
+            unlinkRoleBtn.addEventListener('click', () => {
+                deleteAllPermissionsOfRole(unlinkRoleBtn.attributes['data-name']);
             });
             useSelect2();
         })
